emptystate: move emptystate to shared di
