+++ clients/apps/web/src/components/Customer/CustomerEventsView.tsx
@@ -6,7 +6,7 @@ import { parseAsString, useQueryState } from 'nuqs'
 import { Events } from '../Events/Events'
 import EventSelect from '../Events/EventSelect'
 import MeterSelector from '../Meter/MeterSelector'
-import { EmptyState } from '../CustomerPortal/EmptyState'
+import { EmptyState } from '../Shared/EmptyState'
 import ShortTextOutlined from '@mui/icons-material/ShortTextOutlined'
 
 export const CustomerEventsView = ({
